update the multisig

+++ packages/hardhat-2.14/contracts/MultiSigFactory.sol
@@ -21,7 +21,6 @@ contract MultiSigFactory {
     uint256 indexed signaturesRequired
   );
 
-
   constructor() {}
 
   modifier onlyRegistered() {
@@ -40,31 +39,45 @@ contract MultiSigFactory {
   function create(
     uint256 _chainId,
     address[] memory _owners,
-    uint _signaturesRequired
+    uint8 _signaturesRequired
   ) public payable {
     uint id = numberOfMultiSigs();
 
-    MultiSigWallet multiSig = (new MultiSigWallet){value: msg.value}(_chainId, _owners, _signaturesRequired, address(this));
+    MultiSigWallet multiSig = (new MultiSigWallet){value: msg.value}(
+      _chainId,
+      _owners,
+      _signaturesRequired,
+      address(this)
+    );
     multiSigs.push(multiSig);
     existsMultiSig[address(multiSig)] = true;
 
-    emit Create(id, address(multiSig), msg.sender, _owners, _signaturesRequired);
+    emit Create(
+      id,
+      address(multiSig),
+      msg.sender,
+      _owners,
+      _signaturesRequired
+    );
     emit Owners(address(multiSig), _owners, _signaturesRequired);
   }
 
-  function numberOfMultiSigs() public view returns(uint) {
+  function numberOfMultiSigs() public view returns (uint) {
     return multiSigs.length;
   }
 
-  function getMultiSig(uint256 _index)
+  function getMultiSig(
+    uint256 _index
+  )
     public
     view
-    returns (
-      address multiSigAddress,
-      uint signaturesRequired,
-      uint balance
-    ) {
-      MultiSigWallet multiSig = multiSigs[_index];
-      return (address(multiSig), multiSig.signaturesRequired(), address(multiSig).balance);
-    }
+    returns (address multiSigAddress, uint signaturesRequired, uint balance)
+  {
+    MultiSigWallet multiSig = multiSigs[_index];
+    return (
+      address(multiSig),
+      multiSig.signaturesRequired(),
+      address(multiSig).balance
+    );
+  }
 }
